Slightly better MQTT handling - announce and set clienttimeout
MQTT servers will dc clients who are failing to update with the
value they set themselves.

The library sets this value to 30secs. We update state by default every
5 minutes, so, we always will disconnect. This is just wasteful. All
we need to do is tell the MQTT server we'll be considered disconnected
after (3 x 5mins) in case we miss a cycle due to wifi down, or something
else.

The client also can test it's own connected client status, but that's
really mostly a noop here.