Better implement client keepalive/idle guarding.
We've not properly called .poll() which allows us to send keepalives
and keep our connection up. Ideally we'll go to some actual lower
power state but this at least properly establishes and keeps up
the client MQTT session as until now it would just break and
reconnect every single message. This is much quieter.

The second thing is we can guard against EINVAL from the bme280
because the max/min values for temp and pressure are documented. For
humidity, we can discard 0% and 100% within 0.5% since that is
most likely also an invalid value to begin with.